feat(explorer): notifications as an X-style page (drop the top navbar)
Replace the top-right account/notifications cluster with the Twitter/X pattern:
- Notifications is a left-nav item (below My Profile) with an unread badge over
  the bell icon, routing to a dedicated /notifications page (sticky header +
  list of rows, each linking to its circle).
- The account/profile control moves back to the bottom of the NavSidebar (the
  top-bar profile duplicated the ProfileDrawer's identity).
- Drop TopBar / UserMenu / NotificationsBell.
- useGroupNotifications split into useNotifications (shared query + mark-read)
  and useNotificationsRealtime (the single Ably subscription, mounted in the
  always-present NavSidebar) so the badge, admin panel and join gate update live.
