+++ apps/explorer/src/components/styles/group-join.css
@@ -96,161 +96,3 @@
   color: var(--intent-distrusted, #ef4444);
 }
 
-/* ── Top navbar (sticky header over the main column) ───────── */
-/* First child of `.main-content` — full-bleed across the main column and
-   flush to the top by cancelling its 48px top + 56px side padding; sticks on
-   scroll. Account cluster pinned right. */
-.gj-topbar {
-  position: sticky;
-  top: 0;
-  z-index: 40;
-  display: flex;
-  align-items: center;
-  justify-content: flex-end;
-  gap: 10px;
-  height: 56px;
-  margin: -48px -56px 24px;
-  padding: 0 22px;
-  background: color-mix(in srgb, var(--ds-bg) 85%, transparent);
-  backdrop-filter: blur(10px);
-  -webkit-backdrop-filter: blur(10px);
-  border-bottom: 1px solid var(--ds-border);
-}
-/* Mobile chrome (MobileHeader) owns the account/notifications there. */
-@media (max-width: 900px) {
-  .gj-topbar {
-    display: none;
-  }
-}
-
-/* Notifications bell */
-.gj-bell {
-  position: relative;
-  display: grid;
-  place-items: center;
-  width: 38px;
-  height: 38px;
-  border-radius: 10px;
-  border: 1px solid var(--ds-border);
-  background: var(--ds-card);
-  color: var(--ds-ink-soft);
-  cursor: pointer;
-  transition: border-color 0.13s, color 0.13s;
-  box-shadow: var(--ds-shadow-card);
-}
-.gj-bell:hover {
-  border-color: var(--ds-muted-soft);
-  color: var(--ds-ink);
-}
-.gj-bell-badge {
-  position: absolute;
-  top: -4px;
-  right: -4px;
-  min-width: 16px;
-  height: 16px;
-  padding: 0 4px;
-  border-radius: 999px;
-  background: var(--intent-distrusted, #ef4444);
-  color: #fff;
-  font-family: var(--ds-font-mono, monospace);
-  font-size: 9.5px;
-  font-weight: 700;
-  display: grid;
-  place-items: center;
-  line-height: 1;
-}
-
-/* Notifications dropdown */
-.gj-notif-menu {
-  width: 320px;
-  max-height: 420px;
-  overflow-y: auto;
-  padding: 0;
-}
-.gj-notif-head {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  padding: 11px 13px;
-  border-bottom: 1px solid var(--ds-border-soft);
-  position: sticky;
-  top: 0;
-  background: var(--ds-card);
-}
-.gj-notif-title {
-  font-size: 13px;
-  font-weight: 650;
-  color: var(--ds-ink);
-}
-.gj-notif-readall {
-  font-size: 11px;
-  color: var(--ds-muted);
-  background: none;
-  border: none;
-  cursor: pointer;
-}
-.gj-notif-readall:hover {
-  color: var(--ds-ink);
-}
-.gj-notif-empty {
-  margin: 0;
-  padding: 24px 13px;
-  text-align: center;
-  font-size: 12.5px;
-  color: var(--ds-muted);
-}
-.gj-notif-list {
-  list-style: none;
-  margin: 0;
-  padding: 0;
-}
-.gj-notif-row {
-  position: relative;
-  padding: 11px 13px;
-  border-top: 1px solid var(--ds-border-soft);
-  cursor: pointer;
-  transition: background 0.12s;
-}
-.gj-notif-row:first-child {
-  border-top: none;
-}
-.gj-notif-row:hover {
-  background: var(--ds-bg-subtle);
-}
-.gj-notif-row.is-unread {
-  background: color-mix(in srgb, var(--ds-accent) 7%, transparent);
-}
-.gj-notif-link {
-  display: block;
-  text-decoration: none;
-  color: inherit;
-}
-.gj-notif-row-title {
-  display: block;
-  font-size: 12.5px;
-  font-weight: 600;
-  color: var(--ds-ink);
-}
-.gj-notif-row-msg {
-  display: block;
-  font-size: 11.5px;
-  color: var(--ds-muted);
-  margin-top: 1px;
-  word-break: break-word;
-}
-.gj-notif-row-when {
-  display: block;
-  font-family: var(--ds-font-mono, monospace);
-  font-size: 9.5px;
-  color: var(--ds-muted-soft);
-  margin-top: 4px;
-}
-.gj-notif-dot {
-  position: absolute;
-  top: 13px;
-  right: 11px;
-  width: 7px;
-  height: 7px;
-  border-radius: 50%;
-  background: var(--ds-accent);
-}
